+++ src/yafowil/widget/richtext/__init__.py
@@ -17,6 +17,16 @@ js = [{
     'resource': 'widget.js',
     'order': 22,
 }]
+js_plone4 = [{
+    'group': 'yafowil.widget.richtext.common',
+    'resource': 'widget_p4.js',
+    'order': 20,
+}]
+js_plone5 = [{
+    'group': 'yafowil.widget.richtext.common',
+    'resource': 'widget_p5.js',
+    'order': 20,
+}]
 
 
 @entry_point(order=10)
@@ -24,3 +34,7 @@ def register():
     from yafowil.widget.richtext import widget
     factory.register_theme('default', 'yafowil.widget.richtext',
                            resourcedir, js=js)
+    factory.register_theme('plone4', 'yafowil.widget.richtext',
+                           resourcedir, js=js_plone4)
+    factory.register_theme('plone5', 'yafowil.widget.richtext',
+                           resourcedir, js=js_plone5)